Fix failing tests if the TERM variable is set to unknown
When running the test-suite with TERM=unknown, the tests fail because
the generated output is not "rich".

This happens because python rich will disable color and style if the
TERM variable is set to either "unknown" or "dumb" [1]

This patch monkey-patches the environment to make the test suite
insensitive to the TERM variable.

[1] https://rich.readthedocs.io/en/stable/console.html?highlight=unknown#environment-variables

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>